+++ cargo-rbmt/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* [BREAKING] Remove the debug assertions flag from `test`, users should set `RUSTFLAGS` directly.
+
 ## [0.3.0] - 2026-05-15
 
 * New `generate` command ensures files are generated.