+++ cargo-rbmt/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * Tweak how build and test args are passed down in `test`.
+* Rename the `--lock-file` flag to `--lockfile` to match cargo conventions (e.g. the unstable `--lockfile-path flag`). Still has an alias for backwards compatibility.
 * [BREAKING] Drop the `bench` command in favor of just running `run -- bench`.
 * [BREAKING] Remove the debug assertions flag from `test`, users should set `RUSTFLAGS` directly.